Refactor: 플레이리스트 토글 컴포넌트로 분리
