+++ .env
@@ -1,0 +1,3 @@
+REACT_APP_KAKAO_REST_API_KEY=53ab9d9beaf347ecc3d5779342ef3562
+REACT_APP_KAKAO_REDIRECT_URI=http://localhost:3000/oauth/kakao
+REACT_APP_YOUTUBE_API_KEY=AIzaSyB-WyiV885WZcrvKmYaADQv0qEbgsvTgR8
