+++ src/Pages/Video/AllVideo.js
@@ -1,18 +1,14 @@
-import React, {useCallback, useEffect, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import Header from "../../Components/Header";
 import {
-  AddNewPlyBtn,
-  AutoFrame,
-  Line, MakeNewPlyBtn, NewPlyInput, OneSelectItemWrapper,
+  Line,
   OneVideoWrapper,
-  SortBox, SwitchBtnLabel,
-  ToggleScrollWrapper,
+  SortBox,
   VideoContainer,
   VideoInfo,
   VideoList,
   VideoListWrapper,
   Wrapper,
-  XButton
 } from "../../Style/Video";
 import TV from '../../Assets/TV.png';
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
@@ -23,8 +19,8 @@ import axios from "axios";
 import PlayListIcon from "../../Assets/Playlist_mini.png";
 import {useNavigate} from "react-router-dom";
 import Add_New_Video from "../../Assets/Add_new_video.png";
-import useInput from "../../Hooks/useInput";
 import Pagination from "../../Components/Pagination";
+import PlaylistToggle from "../../Components/PlaylistToggle";
 
 const AllVideo = () => {
   const navigate = useNavigate();
@@ -45,13 +41,12 @@ const AllVideo = () => {
   const [pages, setPages] = useState([1,2,3,4,5]);
   const [page, setPage] = useState(0);  // 현재 페이지
   const [sort, setSort] = useState(""); // 좋아요 순이면 -> likeCount,DESC
-  const [playlistToggleDisplay, setPlaylistToggleDisplay] = useState(false);  // 플레이리스트 모달창 보이기
-  const [playlistList, setPlaylistList] = useState([]); // 받아온 내 플레이리스트 목록
+  const [playListToggleDisplay, setPlayListToggleDisplay] = useState(false);  // 플레이리스트 모달창 보이기
+  const [playListList, setPlayListList] = useState([]); // 받아온 내 플레이리스트 목록
   const [clickedPlyId, setClickedPlyId] = useState(-1); // 클릭한 플레이리스트 아이콘 id
-  // 새로운 플레이리스트
-  const [addNewPly, setAddNewPly] = useState(false);
-  const [newPlyName, onChangeNewPlyName, setNewPlyName] = useInput("");
-  const [newPlyPublic, setNewPlyPublic] = useState(false);
+  /* 등록 처리 필요 */
+  const [playList, setPlayList] = useState([]);
+  const [selectedPlayListId, setSelectedPlayListId] = useState([]);
   // 검색
   const [searchHashtag1, setSearchHashtag1] = useState("");
   const [searchHashtag2, setSearchHashtag2] = useState("");
@@ -100,92 +95,22 @@ const AllVideo = () => {
     axios
         .get(preURL.preURL + `/boards/playlist/user/${1}`)  /*사용자 id*/
         .then((res) => {
-          setPlaylistList(res.data);
+          setPlayListList(res.data);
           console.log("👍내 플레이리스트 조회 성공", res.data);
         })
         .catch((err) => {
           console.log(err);
           console.log("🧨내 플레이리스트 조회 실패", err);
         })
-  },[playlistToggleDisplay]);
+  },[playListToggleDisplay]);
 
   // 플레이리스트에 추가하기 아이콘 클릭
   const onClickAddToPlaylist = (e) => {
     console.log("플레이리스트에 추가", e);
     const clicked = parseInt(e.target.id);
     setClickedPlyId(clicked);
-    setPlaylistToggleDisplay(prev => !prev);
+    setPlayListToggleDisplay(prev => !prev);
   };
-
-  // 플레이리스트 공개/비공개
-  const onClickPublic = (prop) => {
-    const Target = prop.target;
-    const id = Target.id;
-    axios
-        .patch(preURL.preURL + `/boards/playlist/${id}`)
-        .then((res) => {
-          console.log("👍플레이리스트 공개/비공개 수정 성공");
-          if(res.status === 200) {
-            prop.target.parentNode.classList.toggle('active');
-            Target.classList.toggle('active');
-            // console.log(prop.target.parentNode.classList)
-            // console.log(Target);
-            if(Target.innerText === "비공개") Target.innerText = "공개";
-            else Target.innerText = "비공개";
-          }
-          else if(res.status === 403) alert("수정 권한이 없습니다.");
-        })
-        .catch((err) => {
-          console.log("🧨플레이리스트 공개/비공개 수정 실패", err);
-        })
-  };
-
-  // 새 플레이리스트 생성
-  const onClickMakePly = () => {
-    axios
-        .post(preURL.preURL + '/boards/playlist', {
-          title: newPlyName,
-          isPublic: newPlyPublic
-        })
-        .then((res) => {
-          console.log("👍새 플레이리스트 생성 성공", res.data);
-          setNewPlyName("");
-          setNewPlyPublic(false);
-          setAddNewPly(false);
-        })
-        .catch((err) => {
-          console.log("🧨새 플레이리스트 생성 실패", err);
-        })
-  };
-
-  // 플레이리스트 토글 리스트
-  const PlayList = playlistList.map((onePlayList) => {
-    return(
-        <OneSelectItemWrapper>
-          <input
-              type="checkbox"
-              id={onePlayList.id}
-              value={onePlayList.title}
-              // onChange={selectPlayList}
-          />
-          <label>
-            {onePlayList.title}
-          </label>
-          <div>
-            {onePlayList.isPublic
-                ?
-                <SwitchBtnLabel>
-                  <span className="active" id={onePlayList.id} onClick={onClickPublic}>공개</span>
-                </SwitchBtnLabel>
-                :
-                <SwitchBtnLabel>
-                  <span id={onePlayList.id} onClick={onClickPublic}>비공개</span>
-                </SwitchBtnLabel>
-            }
-          </div>
-        </OneSelectItemWrapper>
-    )
-  });
 
   // 최신순 정렬
   const onClickSortNewest = () => {
@@ -228,42 +153,15 @@ const AllVideo = () => {
                     onClick={onClickAddToPlaylist}
                     style={{marginLeft: "4px", cursor: "pointer"}}/>
                 {clickedPlyId === videoId &&   /*클릭한 아이콘과 id가 동일한 모달창에만 적용되도록*/
-                    <AutoFrame display={playlistToggleDisplay} style={{marginTop: "200px"}}>
-                      <XButton onClick={() => {
-                        setPlaylistToggleDisplay(prev => !prev)
-                        setAddNewPly(false);
-                      }}>&times;</XButton>
-                      <span>플레이리스트에 담기</span>
-                      <hr/>
-                      <ToggleScrollWrapper>
-                        {PlayList}
-                      </ToggleScrollWrapper>
-                      <div style={{alignSelf: "center"}}>
-                        {addNewPly
-                            ?
-                            <>
-                              <NewPlyInput type="text" placeholder="플레이리스트 이름" value={newPlyName} onChange={onChangeNewPlyName}/>
-                              <div style={{display: "flex", alignItems: "center", justifyContent: "space-evenly", margin: "10px 0"}}>
-                                {newPlyPublic
-                                    ?
-                                    <SwitchBtnLabel style={{margin: 0}}>
-                                      <span className="active" onClick={() => setNewPlyPublic(prev => !prev)}>공개</span>
-                                    </SwitchBtnLabel>
-                                    :
-                                    <SwitchBtnLabel style={{margin: 0}}>
-                                      <span onClick={() => setNewPlyPublic(prev => !prev)}>비공개</span>
-                                    </SwitchBtnLabel>
-                                }
-                                <MakeNewPlyBtn onClick={onClickMakePly}>만들기</MakeNewPlyBtn>
-                              </div>
-                            </>
-                            :
-                            <AddNewPlyBtn onClick={() => setAddNewPly(prev => !prev)}>
-                              새 플레이리스트 만들기
-                            </AddNewPlyBtn>
-                        }
-                      </div>
-                    </AutoFrame>
+                    <PlaylistToggle
+                        show={playListToggleDisplay}
+                        setShow={setPlayListToggleDisplay}
+                        playListList={playListList}
+                        playList={playList}
+                        setPlayList={setPlayList}
+                        selectedPlayListId={selectedPlayListId}
+                        setSelectedPlayListId={setSelectedPlayListId}
+                    />
                 }
               </div>
             </VideoInfo>
